deleteAccount to command
