+++ packages/hedera-wallet-snap/packages/snap/src/client/commands/DeleteAccountCommand.ts
@@ -1,0 +1,58 @@
+import { AccountDeleteTransaction, Client } from '@hashgraph/sdk';
+import { TxReceipt } from '../../types/hedera';
+import { Utils } from '../../utils/Utils';
+import { CryptoUtils } from '../../utils/CryptoUtils';
+
+export class DeleteAccountCommand {
+  readonly #transferAccountId: string;
+
+  constructor(transferAccountId: string) {
+    this.#transferAccountId = transferAccountId;
+  }
+
+  public async execute(client: Client): Promise<TxReceipt> {
+    const transaction = new AccountDeleteTransaction()
+      .setAccountId(
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        client.operatorAccountId!,
+      )
+      .setTransferAccountId(this.#transferAccountId)
+      .freezeWith(client);
+
+    const txResponse = await transaction.execute(client);
+
+    const receipt = await txResponse.getReceipt(client);
+
+    let newExchangeRate;
+    if (receipt.exchangeRate) {
+      newExchangeRate = {
+        ...receipt.exchangeRate,
+        expirationTime: Utils.timestampToString(
+          receipt.exchangeRate.expirationTime,
+        ),
+      };
+    }
+
+    return {
+      status: receipt.status.toString(),
+      accountId: receipt.accountId ? receipt.accountId.toString() : '',
+      fileId: receipt.fileId ? receipt.fileId : '',
+      contractId: receipt.contractId ? receipt.contractId : '',
+      topicId: receipt.topicId ? receipt.topicId : '',
+      tokenId: receipt.tokenId ? receipt.tokenId : '',
+      scheduleId: receipt.scheduleId ? receipt.scheduleId : '',
+      exchangeRate: newExchangeRate,
+      topicSequenceNumber: receipt.topicSequenceNumber
+        ? String(receipt.topicSequenceNumber)
+        : '',
+      topicRunningHash: CryptoUtils.uint8ArrayToHex(receipt.topicRunningHash),
+      totalSupply: receipt.totalSupply ? String(receipt.totalSupply) : '',
+      scheduledTransactionId: receipt.scheduledTransactionId
+        ? receipt.scheduledTransactionId.toString()
+        : '',
+      serials: JSON.parse(JSON.stringify(receipt.serials)),
+      duplicates: JSON.parse(JSON.stringify(receipt.duplicates)),
+      children: JSON.parse(JSON.stringify(receipt.children)),
+    } as TxReceipt;
+  }
+}
